Module	Files	Key changes
types	types.rs	All core structs (AgentState, Market, Estimate, Position, TradeReceipt, etc.)
config	config.rs	RiskConfig, AgentConfig financial fields via serde-float
strategy	edge.rs, kelly.rs, risk.rs, mod.rs	Thresholds, sizing, exposure limits
platforms	polymarket.rs, manifold.rs, metaculus.rs, mod.rs	Trait signatures, boundary conversion
llm	anthropic.rs, openai.rs, openrouter.rs, mod.rs	cost_per_call(), Estimate construction
engine	executor.rs, accountant.rs, scanner.rs, enricher.rs	CycleCosts, reports, filtering
data	weather.rs, sports.rs, economics.rs, news.rs, mod.rs	DataProvider cost_per_call()
dashboard	routes.rs, mod.rs	Response structs stay f64, convert via .to_f64()
backtest	runner.rs	Money → Decimal, stats (Sharpe, Brier) stay f64
storage	mod.rs	tracing Display formatting
main	main.rs	Config → strategy wiring, format strings

+++ src/data/mod.rs
@@ -10,6 +10,7 @@ pub mod news;
 
 use anyhow::Result;
 use async_trait::async_trait;
+use rust_decimal::Decimal;
 
 use crate::types::{DataContext, Market, MarketCategory};
 
@@ -26,5 +27,5 @@ pub trait DataProvider: Send + Sync {
     async fn fetch_context(&self, market: &Market) -> Result<DataContext>;
 
     /// Cost per API call in USD (for survival accounting).
-    fn cost_per_call(&self) -> f64;
+    fn cost_per_call(&self) -> Decimal;
 }
